chapter_12: add initial alien invasion game
This commit adds the main game structure, settings,
ship class and the ship image to be the main character.

At this point, the game opens a new window with
a background color, a image of the main ship and
it's possible to move it left and right using
the arrow keys.

Signed-off-by: ltsuda <leonardo.tsuda@icloud.com>